Update codebase for SS4
