+++ composer.json
@@ -1,6 +1,6 @@
 {
   "name": "guttmann/silverstripe-menumanager-subsites",
-  "type": "silverstripe-module",
+  "type": "silverstripe-vendormodule",
   "license": "MIT",
   "authors": [
     {
@@ -9,14 +9,18 @@
     }
   ],
   "require": {
-    "php": ">=5.3.2",
-    "composer/installers": "*",
-    "silverstripe/framework": "~3.1",
-    "silverstripe/cms": "~3.1",
-    "heyday/silverstripe-menumanager": "~2.0",
+    "silverstripe/cms": "^4",
+    "heyday/silverstripe-menumanager": "^3",
     "silverstripe/subsites": "~1.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "Guttmann\\SilverStripe\\": "src/"
+    }
   },
   "extra": {
     "installer-name": "menumanager_subsites"
-  }
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true
 }
