+++ src/SubsiteMenuManagerTemplateProvider.php
@@ -2,12 +2,12 @@
 
 namespace Guttmann\SilverStripe;
 
-use MenuSet;
-use Subsite;
+use Heyday\MenuManager\MenuSet;
+use SilverStripe\Subsites\State\SubsiteState;
+use SilverStripe\View\TemplateGlobalProvider;
 
-class SubsiteMenuManagerTemplateProvider implements \TemplateGlobalProvider
+class SubsiteMenuManagerTemplateProvider implements TemplateGlobalProvider
 {
-
     public static function get_template_global_variables()
     {
         return array(
@@ -19,7 +19,7 @@ class SubsiteMenuManagerTemplateProvider implements \TemplateGlobalProvider
     {
         return MenuSet::get()->filter(array(
             'Name' => $name,
-            'SubsiteID' => Subsite::currentSubsiteID()
+            'SubsiteID' => SubsiteState::singleton()->getSubsiteId()
         ))->First();
     }
